[refactor] ModalWrapper의 Overlay를 Layout에 있는 것으로 사용
- 기존에는 따로 Overlay를 만들어서 사용하고 있었으나 Overlay Layout컴포넌트와 하는 역할이 동일함으로 그것을 사용하였다.
- Overlay클릭시 발생하는 함수 이름이 부적절하다고 판단하여 수정하였다
